pre-create all download sub-directories
this fixes download directory creation which breaks when done on
the fly inside the download loop.

from the error that results without this change, it seems that the
`await mkdirp()` call doesn't *actually* await?

i have no idea, i'm not a nodejs dev!

regardless, pre-creating all the required directories is obviously the
best thing to do here since we know what we need from the start.